fix `delete` method

+++ mango-orm/src/models/db_query_api/paladins.rs
@@ -1327,14 +1327,19 @@ pub trait QPaladins: ToModel + CachingModel {
                                                 fs::remove_file(path)?;
                                             }
                                             // Remove thumbnails.
-                                            let size_names: [&str; 4] = ["lg", "md", "sm", "xs"];
-                                            for size_name in size_names.iter() {
-                                                let key_name = format!("{}_{}", "path", size_name);
-                                                let path = field_file.get_str(key_name.as_str())?;
-                                                if path.is_empty() {
-                                                    let path = Path::new(path);
-                                                    if path.exists() {
-                                                        fs::remove_file(path)?;
+                                            if widget_name == "inputImage" {
+                                                let size_names: [&str; 4] =
+                                                    ["lg", "md", "sm", "xs"];
+                                                for size_name in size_names.iter() {
+                                                    let key_name =
+                                                        format!("{}_{}", "path", size_name);
+                                                    let path =
+                                                        field_file.get_str(key_name.as_str())?;
+                                                    if path.is_empty() {
+                                                        let path = Path::new(path);
+                                                        if path.exists() {
+                                                            fs::remove_file(path)?;
+                                                        }
                                                     }
                                                 }
                                             }
